// web/src/lib/api/generated/model/newUserResponse.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Spark API
 * Spark API
 * OpenAPI spec version: 0.1.0
 */

export interface NewUserResponse {
  userId?: number;
  name?: string;
  /**
     * プロフィールアイコン画像の URL
     * @nullable
     */
  iconUrl?: string | null;
}