// web/src/lib/api/generated/model/pickupUserResponse.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Spark API
 * Spark API
 * OpenAPI spec version: 0.1.0
 */
import type { ThingResponse } from './thingResponse';

export interface PickupUserResponse {
  userId?: number;
  name?: string;
  /** @nullable */
  bio?: string | null;
  /**
     * プロフィールアイコン画像の URL
     * @nullable
     */
  iconUrl?: string | null;
  /** ログインユーザーと一致するタグの一覧 */
  matchedTags?: ThingResponse[];
  /** ログインユーザーと一致しないタグの一覧 */
  unmatchedTags?: ThingResponse[];
}